+++ src/main/kotlin/com/tezos/lang/michelson/editor/liveTemplate/MichelsonToplevelContext.kt
@@ -1,13 +1,12 @@
 package com.tezos.lang.michelson.editor.liveTemplate
 
 import com.intellij.codeInsight.template.TemplateContextType
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiErrorElement
 import com.intellij.psi.PsiFile
+import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.util.PsiUtil
 import com.tezos.lang.michelson.lang.MichelsonLanguage
-import com.tezos.lang.michelson.psi.MichelsonPsiFile
-import com.tezos.lang.michelson.psi.PsiContract
+import com.tezos.lang.michelson.psi.PsiCodeSection
+import com.tezos.lang.michelson.psi.PsiTypeSection
 
 /**
  * @author jansorg
@@ -18,6 +17,9 @@ class MichelsonToplevelContext : TemplateContextType("MICHELSON_TOPLEVEL", "Mich
             return false
         }
 
-        return true
+        // show templates which are not in any of the well known sections parameter, storage, or code.
+        // this accepts unknown sections which accept errors
+        val psi = file.findElementAt(offset)
+        return PsiTreeUtil.findFirstParent(psi) { it is PsiCodeSection || it is PsiTypeSection } == null
     }
 }